% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_indic.R
\encoding{UTF-8}
\name{carte_indic}
\alias{carte_indic}
\title{Carte pour rpls}
\usage{
carte_indic(.data = indicateurs_rpls, indicateur, zoom_reg = F,
  sousensemble = "Ensemble du parc", legend = F, titre = NULL,
  soustitre = NULL, basdepage = NULL, filtre_zero = F,
  variable = Pourcent, na_recode = "Pas de logements", box = bbox)
}
\arguments{
\item{.data}{le dataframe de départ}

\item{indicateur}{indicateur à cartographier}

\item{zoom_reg}{booléen T si on veut la carte régional, F pour la carte national}

\item{sousensemble}{sous ensemble du parc à cartographier}

\item{legend}{T si on veut une légende}

\item{titre}{titre du graphique}

\item{soustitre}{sous-titre du graphique}

\item{basdepage}{bas de page du graphique}

\item{filtre_zero}{T si on veut ne pas utiliser les valeurs à zéro pour définir les classes de valeur}

\item{variable}{variable à valoriser : poucentage ou valeur absolue}

\item{na_recode}{libellé pour les valeur à NA de l'indicateur}

\item{box}{vecteur des coordonnées du territore sur lequel zoomer}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Carte pour rpls
}
