% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_indic_croissance.R
\name{carte_indic_croissance}
\alias{carte_indic_croissance}
\title{Carte pour rpls}
\usage{
carte_indic_croissance(zoom_reg = F, box = bbox, g = guide)
}
\arguments{
\item{zoom_reg}{booléen T si on veut la carte régional, F pour la carte national}

\item{box}{vecteur des coordonnées du territore sur lequel zoomer}

\item{g}{guide}

\item{indicateur}{indicateur à cartographier}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Carte pour rpls
}
