% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\encoding{UTF-8}
\name{carte_synthese}
\alias{carte_synthese}
\title{Carte en rond proportionnel donnant sur une région le taux de logements sociaux et le nombre de logements sociaux}
\usage{
carte_synthese(.data = indicateurs_rpls, titre = NULL,
  soustitre = NULL, basdepage = NULL, box = bbox)
}
\arguments{
\item{.data}{le dataframe de départ}

\item{titre}{titre du graphique}

\item{soustitre}{sous-titre du graphique}

\item{basdepage}{bas de page du graphique}

\item{box}{vecteur des coordonnées du territore sur lequel zoomer}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Carte en rond proportionnel donnant sur une région le taux de logements sociaux et le nombre de logements sociaux
}
