% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_taux_evolution.R
\encoding{UTF-8}
\name{carte_taux_evolution}
\alias{carte_taux_evolution}
\title{Carte du taux d'évolution}
\usage{
carte_taux_evolution(.data = indicateurs_rpls, zoom_reg = F,
  na_recode = "Pas de parc social", box = bbox)
}
\arguments{
\item{.data}{le dataframe en entrée}

\item{zoom_reg}{booléen T si on veut la carte régional, F pour la carte national}

\item{na_recode}{chaine de caractères, le libellé qui s'affichera sur la carte pour le valeurs manquantes}

\item{box}{vecteur des coordonnées du territoire sur lequel zoomer}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Carte du taux d'évolution
}
