% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeur_carte.R
\encoding{UTF-8}
\name{composeur_carte}
\alias{composeur_carte}
\title{Composeur de carte pour rpls}
\usage{
composeur_carte(.data = indicateurs_rpls, indicateur, titre = NULL,
  soustitre = NULL, basdepage = NULL, filtre_zero = F,
  variable = Pourcent, na_recode = "Pas de logements",
  parc_recent = T, box = bbox, g = guide)
}
\arguments{
\item{.data}{le dataframe de départ}

\item{indicateur}{indicateur à cartographier}

\item{titre}{titre du graphique}

\item{soustitre}{sous-titre du graphique}

\item{basdepage}{bas de page du graphique}

\item{filtre_zero}{T si on veut ne pas utiliser les valeurs à zéro pour définir les classes de valeur}

\item{variable}{variable à valoriser : poucentage ou valeur absolue}

\item{na_recode}{libellé pour les valeur à NA de l'indicateur}

\item{parc_recent}{booléen : T si on veut la carte sur le parc de moins de 5 ans}

\item{box}{vecteur des coordonnées du territore sur lequel zoomer}

\item{g}{guide}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Composeur de carte pour rpls
}
