% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls_2018.R
\encoding{UTF-8}
\name{creer_indicateurs_rpls_2018}
\alias{creer_indicateurs_rpls_2018}
\title{Creation des indicateurs pour la publication}
\usage{
creer_indicateurs_rpls_2018(.data, sous_ensemble = "Ensemble du parc")
}
\arguments{
\item{.data}{le fichier en entree}

\item{sous_ensemble}{le sous ensemble du parc pris en compte}
}
\value{
la fonction renvoie un dataframe
}
\description{
Creation des indicateurs pour la publication
}
