% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls_2018.R
\encoding{UTF-8}
\name{creer_indicateurs_rpls_2018_publication}
\alias{creer_indicateurs_rpls_2018_publication}
\title{Creation des indicateurs pour la publication sur les 3 sous ensembles}
\usage{
creer_indicateurs_rpls_2018_publication(.data)
}
\arguments{
\item{.data}{le fichier en entree}
}
\value{
la fonction renvoie un dataframe
}
\description{
Creation des indicateurs pour la publication sur les 3 sous ensembles
}
