% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{euro_m2_format}
\alias{euro_m2_format}
\title{Formatage d'un nombre en montant en euros}
\usage{
euro_m2_format(x)
}
\arguments{
\item{x}{un nombre}
}
\description{
Formatage d'un nombre en montant en euros
}
\examples{
nombre_format(400.5)
}
