% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{euro_m2_format_ggplot}
\alias{euro_m2_format_ggplot}
\title{Formatage d'un nombre en montant en euros sans le m2 en exposant}
\usage{
euro_m2_format_ggplot(x)
}
\arguments{
\item{x}{un nombre}
}
\value{
chaine de caractère
}
\description{
Formatage d'un nombre en montant en euros sans le m2 en exposant
}
\examples{
nombre_format(400.5)
}
