% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{evo_parc_hlm}
\alias{evo_parc_hlm}
\title{chiffre clef - evolution du parc hlm en pourcent}
\usage{
evo_parc_hlm(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe ou se trouve les indicateurs}
}
\value{
un chaine de charactère
}
\description{
chiffre clef - evolution du parc hlm en pourcent
}
