% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_repartition_par_date.R
\name{graph_repartition_par_date}
\alias{graph_repartition_par_date}
\title{Graphique comparaison par date de construction pour rpls}
\usage{
graph_repartition_par_date(indicateur, titre = NULL,
  soustitre = "Par date de construction", basdepage = NULL,
  date_debut = 1950)
}
\arguments{
\item{indicateur}{indicateur à cartographier}

\item{titre}{titre du graphique}

\item{soustitre}{sous-titre du graphique}

\item{basdepage}{bas de page du graphique}

\item{date_debut}{date de début pour l'analyse}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Graphique comparaison par date de construction pour rpls
}
