% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphique_comparaison_parc_recent.R
\name{graphique_comparaison_parc_recent}
\alias{graphique_comparaison_parc_recent}
\title{Graphique comparaison par récent pour rpls}
\usage{
graphique_comparaison_parc_recent(.data = df, indicateur,
  variable = Pourcent, title = "", caption = "")
}
\arguments{
\item{.data}{le dataframe de départ}

\item{indicateur}{indicateur à cartographier}

\item{variable}{variable à valoriser : poucentage ou valeur absolue}

\item{title}{titre du graphique}

\item{caption}{bas de page du graphique}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Graphique comparaison par récent pour rpls
}
