% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{graphique_synthese_collectif}
\alias{graphique_synthese_collectif}
\title{waffle sur le taux de collectif dans le parc social}
\usage{
graphique_synthese_collectif(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe de départ}
}
\value{
un ggplot waffle chart
}
\description{
waffle sur le taux de collectif dans le parc social
}
