% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{graphique_synthese_loyer}
\alias{graphique_synthese_loyer}
\title{graphique barre qui compare le loyer au m2 sur une région sur sur la France entière}
\usage{
graphique_synthese_loyer(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe de départ}
}
\value{
un ggplot graphique barre
}
\description{
graphique barre qui compare le loyer au m2 sur une région sur sur la France entière
}
