% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{graphique_synthese_parc_epci}
\alias{graphique_synthese_parc_epci}
\title{graphique barre donnant le nombre de logements sociaux sur les principaux epci}
\usage{
graphique_synthese_parc_epci(.data = indicateurs_rpls, titre = NULL,
  soustitre = NULL, basdepage = NULL, nombre_epci = 6)
}
\arguments{
\item{.data}{le dataframe de départ}

\item{titre}{titre du graphique}

\item{soustitre}{sous-titre du graphique}

\item{basdepage}{bas de page du graphique}

\item{nombre_epci}{nombre d'epci à afficher}
}
\value{
un ggplot
}
\description{
graphique barre donnant le nombre de logements sociaux sur les principaux epci
}
