% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{graphique_synthese_taux_hlm}
\alias{graphique_synthese_taux_hlm}
\title{waffle sur le taux de hlm sur une région}
\usage{
graphique_synthese_taux_hlm(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe de départ}
}
\value{
un ggplot waffle chart
}
\description{
waffle sur le taux de hlm sur une région
}
