% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicateurs_rpls_2019}
\alias{indicateurs_rpls_2019}
\title{Indicateurs calculés sur RPLS 2019 au territoire (communes, EPCI, Départements, Régions)}
\format{a \code{tbl_df} containing the following fields:
\describe{
\item{TypeZone}{Communes, Epci, Départements, Régions ou France}
\item{Zone}{Libellé du territoire}
\item{CodeZone}{Code du territoire}
\item{Variable}{Variable à partir de laquelle a été construit l'indicateur}
\item{Indicateur}{Libellé de l'indicateur}
\item{Valeur}{Valeur de l'indicateur}
\item{Pourcent}{Pourcentage de la valeur de l'indicateur par rapport au nombre de logements sociaux total}
\item{SousEnsemble}{Ensemble du parc ou parc de moins de 5 ans}
\item{Valeur_2018}{Valeur de l'indicateur en 2018}
\item{Pourcent}{Pourcentage de la valeur de l'indicateur par rapport au nombre de logements sociaux total en 2018}
}}
\source{
RPLS 2018.
}
\description{
Indicateurs précalculés au territoire
}
\seealso{
\link{rrrpls}
}
