% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legende_carte_indic.R
\name{legende_carte_indic}
\alias{legende_carte_indic}
\title{Légende pour carte pour rpls}
\usage{
legende_carte_indic(indicateur, filtre_zero = F, zoom_reg = F,
  variable = Pourcent, na_recode = "Pas de logements")
}
\arguments{
\item{indicateur}{indicateur à cartographier}

\item{filtre_zero}{T si on veut ne pas utiliser les valeurs à zéro pour définir les classes de valeur}

\item{zoom_reg}{booléen T si on veut la carte régional, F pour la carte national}

\item{variable}{variable à valoriser : poucentage ou valeur absolue}

\item{na_recode}{libellé pour les valeur à NA de l'indicateur}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Légende pour carte pour rpls
}
