% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls.R
\name{lire_rep_rpls}
\alias{lire_rep_rpls}
\title{Importer des fichers rpls csv en dataframe}
\usage{
lire_rep_rpls(dir, annee, geoloc = F, as_sf = F)
}
\arguments{
\item{annee}{le millésime du fichier}

\item{geoloc}{booléen TRUE si on souhaite importer la versin géolocalisée du millésime}

\item{as_sf}{booléen TRUE si on souhaite convertir le dataframe en spatial dataframe (pour les données géolocalisées). Par défaut, utilisation de la projection Lambert93.}

\item{file}{le répertoire ou se trouve les fichiers csv}
}
\value{
la fonction renvoie un dataframe ou un spatial dataframe qui contient l'ensemble des données contenues dans les csv
}
\description{
Importer des fichers rpls csv en dataframe
}
