% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls_2018.R
\encoding{UTF-8}
\name{lire_rep_rpls_loi_2018}
\alias{lire_rep_rpls_loi_2018}
\title{Importer des fichers rpls csv 2018 en dataframe}
\usage{
lire_rep_rpls_loi_2018(dir, passer_au_cog_a_jour = TRUE)
}
\arguments{
\item{dir}{le repertoire ou se trouve les fichiers csv}

\item{passer_au_cog_a_jour}{booleen TRUE si on souhaite passer les données au millesime 2018 du COG}
}
\value{
la fonction renvoie un dataframe qui contient l'ensemble des donnees contenues dans les csv
}
\description{
Importer des fichers rpls csv 2018 en dataframe
}
