% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls.R
\name{lire_rpls}
\alias{lire_rpls}
\title{Importer une ficher rpls csv en Rdata}
\usage{
lire_rpls(file, annee, geoloc = F, as_sf = F)
}
\arguments{
\item{file}{le fichier csv}

\item{annee}{le millésime du fichier}

\item{geoloc}{booléen TRUE si on souhaite importer la versin géolocalisée du millésime}

\item{as_sf}{booléen TRUE si on souhaite convertir le dataframe en spatial dataframe (pour les données géolocalisées)}
}
\value{
la fonction renvoie un dataframe ou un spatial dataframe
}
\description{
Importer une ficher rpls csv en Rdata
}
