% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls_2017.R
\name{lire_rpls_2017_geoloc}
\alias{lire_rpls_2017_geoloc}
\title{Importer une ficher rpls 2017 geolocalisé csv en Rdata}
\usage{
lire_rpls_2017_geoloc(file, as_sf = F)
}
\arguments{
\item{file}{le fichier csv}

\item{as_sf}{booléen TRUE si on souhaite convertir le dataframe en spatial dataframe (pour les données géolocalisées)}
}
\value{
la fonction renvoie un dataframe ou un spatial dataframe
}
\description{
Importer une ficher rpls 2017 geolocalisé csv en Rdata
}
