% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls_2018.R
\encoding{UTF-8}
\name{lire_rpls_loi_2018}
\alias{lire_rpls_loi_2018}
\title{Importer un ficher rpls 2018 loi csv en Rdata}
\usage{
lire_rpls_loi_2018(file, passer_au_cog_a_jour = TRUE)
}
\arguments{
\item{file}{le fichier csv}

\item{passer_au_cog_a_jour}{booleen TRUE si on souhaite passer les données au millesime 2018 du COG}
}
\value{
la fonction renvoie un dataframe
}
\description{
Importer un ficher rpls 2018 loi csv en Rdata
}
