% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{loyer_m2}
\alias{loyer_m2}
\title{chiffre clef - loyer au m2 moyen dans le parc social}
\usage{
loyer_m2(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe ou se trouve les indicateurs}
}
\value{
un chaine de charactère
}
\description{
chiffre clef - loyer au m2 moyen dans le parc social
}
