% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{nombre_format}
\alias{nombre_format}
\title{Formatage d'un nombre avec les conventions françaises}
\usage{
nombre_format(x)
}
\arguments{
\item{x}{un nombre}
}
\description{
Formatage d'un nombre avec les conventions françaises
}
\examples{
nombre_format(4)
}
