% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{parc_hlm}
\alias{parc_hlm}
\title{chiffre clef - nombre total de logements}
\usage{
parc_hlm(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe ou se trouve les indicateurs}
}
\value{
un chaine de charactère
}
\description{
chiffre clef - nombre total de logements
}
