% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{pourcent_format}
\alias{pourcent_format}
\title{Formatage d'un nombre en pourcentage français}
\usage{
pourcent_format(x)
}
\arguments{
\item{x}{un nombre}
}
\value{
chaine de caractère
}
\description{
Formatage d'un nombre en pourcentage français
}
\examples{
nombre_format(10.2)
}
