% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{pourcent_parc_hlm}
\alias{pourcent_parc_hlm}
\title{chiffre clef - pourcentage du logements hlm dans le parc total}
\usage{
pourcent_parc_hlm(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe ou se trouve les indicateurs}
}
\value{
un chaine de charactère
}
\description{
chiffre clef - pourcentage du logements hlm dans le parc total
}
