% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_synthese.R
\name{pourcent_qpv}
\alias{pourcent_qpv}
\title{chiffre clef - pourcentage de logements en QPV dans le parc social}
\usage{
pourcent_qpv(.data = indicateurs_rpls)
}
\arguments{
\item{.data}{le dataframe ou se trouve les indicateurs}
}
\value{
un chaine de charactère
}
\description{
chiffre clef - pourcentage de logements en QPV dans le parc social
}
