% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpls_par_date_2018}
\alias{rpls_par_date_2018}
\title{Données rpls aggrégés par date et communes}
\format{un \code{tbl_df} contenant les champs suivants:
\describe{
\item{DEPCOM}{Code commune}
\item{construct_red}{date de construction}
\item{dpeenergie_red}{dpe energie}
\item{dpeserre_red}{dpe GES}
\item{nbpiece_red}{Nombre de piece aggreges a 5 et +}
\item{typeconst_red}{collectif ou individuel}
\item{qpv_red}{en ou hors qpv}
\item{n}{nombre de logements}
}}
\source{
RPLS 2018.
}
\description{
Données rpls aggrégés par date et communes
}
\seealso{
\link{rrrpls}
}
