% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpls_par_date_2019}
\alias{rpls_par_date_2019}
\title{Données rpls aggrégés par date et communes}
\format{un \code{tbl_df} contenant les champs suivants:
\describe{
\item{DEPCOM}{Code commune}
\item{construct}{date de construction}
\item{dpeenergie}{dpe energie}
\item{dpeserre}{dpe GES}
\item{nbpiece}{Nombre de piece aggreges a 5 et +}
\item{typeconst}{collectif ou individuel}
\item{qpv}{en ou hors qpv}
\item{n}{nombre de logements}
}}
\source{
RPLS 2019.
}
\description{
Données rpls aggrégés par date et communes
}
\seealso{
\link{rrrpls}
}
