% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls_2018.R
\encoding{UTF-8}
\name{selectionner_rpls_loi_2018}
\alias{selectionner_rpls_loi_2018}
\title{Mettre en forme les donées rpls}
\usage{
selectionner_rpls_loi_2018(.data)
}
\arguments{
\item{.data}{le dataframe des données rpls loi 2018}
}
\value{
la fonction renvoie un dataframe qui contient une sélection de données
}
\description{
Mettre en forme les donées rpls
}
