% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_indic_croissance.R
\encoding{UTF-8}
\name{carte_indic_croissance}
\alias{carte_indic_croissance}
\title{Carte pour rpls}
\usage{
carte_indic_croissance(.data = indicateurs_rpls, zoom_reg = F,
  na_recode = "Pas de parc récent", box = bbox)
}
\arguments{
\item{.data}{le dataframe en entrée}

\item{zoom_reg}{booléen T si on veut la carte régional, F pour la carte national}

\item{box}{vecteur des coordonnées du territoire sur lequel zoomer}
}
\value{
la fonction renvoie un graphique ggplot2
}
\description{
Carte pour rpls
}
