% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_rpls.R
\name{lire_rep_rpls}
\alias{lire_rep_rpls}
\title{Importer des fichers rpls csv en dataframe}
\usage{
lire_rep_rpls(dir, annee, geoloc = F, as_sf = F)
}
\arguments{
\item{dir}{le repertoire ou se trouve les fichiers csv}

\item{annee}{le millesime du fichier}

\item{geoloc}{booleen TRUE si on souhaite importer la versin geolocalisee du millesime}

\item{as_sf}{booleen TRUE si on souhaite convertir le dataframe en spatial dataframe (pour les donnees geolocalisees). Par defaut, utilisation de la projection Lambert93.}
}
\value{
la fonction renvoie un dataframe ou un spatial dataframe qui contient l'ensemble des donnees contenues dans les csv
}
\description{
Importer des fichers rpls csv en dataframe
}
